# contracts/openapi/control-plane-v1.yaml
openapi: 3.1.0
info:
  title: Molejo Control Plane API
  version: 1.0.0
  description: Pre-alpha control-plane contract for the local beta workspace.
servers:
  - url: /
paths:
...
  /api/v1/workspaces/{workspaceId}/github/installations/connect:
    parameters: [{ $ref: '#/components/parameters/WorkspaceId' }]
    post:
      operationId: connectGitHubInstallation
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Short-lived GitHub App installation URL
          content:
            application/json:
              schema:
                type: object
                required: [authorizationUrl]
                properties:
                  authorizationUrl: { type: string, format: uri }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/v1/github/installations/callback:
    get:
      operationId: completeGitHubInstallation
      parameters:
        - name: installation_id
          in: query
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
        - name: state
          in: query
          required: true
          schema: { type: string, minLength: 32, maxLength: 256 }
      responses:
        '302': { description: Continue with GitHub user authorization }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/v1/github/callback:
    get:
      operationId: completeGitHubAuthorization
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 512 }
        - name: state
          in: query
          required: true
          schema: { type: string, minLength: 32, maxLength: 256 }
      responses:
        '302': { description: Return to the Console }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /api/v1/workspaces/{workspaceId}/github/installations:
    parameters: [{ $ref: '#/components/parameters/WorkspaceId' }]
    get:
      operationId: listGitHubInstallations
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: GitHub App installations connected to the Workspace
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/GitHubInstallation' } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/workspaces/{workspaceId}/github/installations/{githubInstallationId}:
    parameters:
      - { $ref: '#/components/parameters/WorkspaceId' }
      - { $ref: '#/components/parameters/GitHubInstallationId' }
    delete:
      operationId: disconnectGitHubInstallation
      security: [{ sessionCookie: [] }]
      responses:
        '204': { description: GitHub App installation disconnected and uninstalled }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/v1/workspaces/{workspaceId}/github/installations/{githubInstallationId}/repositories:
    parameters:
      - { $ref: '#/components/parameters/WorkspaceId' }
      - { $ref: '#/components/parameters/GitHubInstallationId' }
    get:
      operationId: listGitHubRepositories
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Repositories accessible to the installation
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/GitHubRepository' } }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/v1/workspaces/{workspaceId}/projects/{projectId}/apps/{appId}/source:
    parameters:
      - { $ref: '#/components/parameters/WorkspaceId' }
      - { $ref: '#/components/parameters/ProjectId' }
      - { $ref: '#/components/parameters/AppId' }
    get:
      operationId: getAppSource
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Optional GitHub source configured for the App
          content:
            application/json:
              schema:
                type: object
                required: [source]
                properties:
                  source:
                    oneOf:
                      - { $ref: '#/components/schemas/GitHubSource' }
                      - { type: 'null' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: setAppSource
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GitHubSourceInput' }
      responses:
        '200':
          description: GitHub source selected for the App
          content: { application/json: { schema: { $ref: '#/components/schemas/GitHubSource' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
    delete:
      operationId: clearAppSource
      security: [{ sessionCookie: [] }]
      responses:
        '204': { description: GitHub source removed from the App }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: fruto_session
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, pattern: '^ws-[a-z2-7]{20}$' }
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, pattern: '^prj-[a-z2-7]{20}$' }
...
    AppId:
      name: appId
      in: path
      required: true
      schema: { type: string, pattern: '^app-[a-z2-7]{20}$' }
...
    GitHubInstallationId:
      name: githubInstallationId
      in: path
      required: true
      schema: { type: string, pattern: '^ghi-[a-z2-7]{20}$' }
...
  responses:
...
    Forbidden:
      description: Authenticated user cannot perform this action
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Version or idempotency conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unavailable:
      description: External integration is not configured or unavailable
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
...
    GitHubInstallation:
      type: object
      required: [id, accountLogin, accountType, repositorySelection, createdAt]
      additionalProperties: false
      properties:
        id: { type: string, pattern: '^ghi-[a-z2-7]{20}$' }
        accountLogin: { type: string, minLength: 1 }
        accountType: { type: string, enum: [User, Organization, Enterprise] }
        repositorySelection: { type: string, enum: [all, selected] }
        createdAt: { type: string, format: date-time }
    GitHubRepository:
      type: object
      required: [id, name, fullName, private, defaultBranch]
      additionalProperties: false
      properties:
        id: { type: string, pattern: '^[1-9][0-9]*$' }
        name: { type: string, minLength: 1 }
        fullName: { type: string, minLength: 3 }
        private: { type: boolean }
        defaultBranch: { type: string, minLength: 1 }
    GitHubSourceInput:
      type: object
      required: [installationId, repositoryId]
      additionalProperties: false
      properties:
        installationId: { type: string, pattern: '^ghi-[a-z2-7]{20}$' }
        repositoryId: { type: string, pattern: '^[1-9][0-9]*$' }
    GitHubSource:
      type: object
      required: [installationId, repository, connectedAt]
      additionalProperties: false
      properties:
        installationId: { type: string, pattern: '^ghi-[a-z2-7]{20}$' }
        repository: { $ref: '#/components/schemas/GitHubRepository' }
        connectedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }